# apps/tickets-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tickets server",
    "version": "0.0.0"
  },
  "paths": {
    "/api/chat": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "threadId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "messages": {
                    "type": "array",
                    "items": {}
                  },
                  "confirmationProposalId": {
                    "type": "string"
                  }
                },
                "required": [
                  "threadId",
                  "messages"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "tenant-a",
                "tenant-b"
              ]
            },
            "in": "header",
            "name": "x-tenant-id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "AI SDK UI message stream",
            "headers": {
              "x-vercel-ai-ui-message-stream": {
                "required": true,
                "schema": {
                  "const": "v1"
                }
              }
            },
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "Unknown or missing tenant.",
                        "Invalid JSON.",
                        "Missing thread."
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "Tenant does not match the thread."
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/decisions": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "proposalId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "decision": {
                    "type": "string",
                    "enum": [
                      "approve",
                      "reject"
                    ]
                  }
                },
                "required": [
                  "proposalId",
                  "decision"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "tenant-a",
                "tenant-b"
              ]
            },
            "in": "header",
            "name": "x-tenant-id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "outcome": {
                      "type": "string",
                      "enum": [
                        "applied",
                        "rejected"
                      ]
                    }
                  },
                  "required": [
                    "outcome"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "Unknown or missing tenant.",
                        "Invalid JSON.",
                        "Missing proposal.",
                        "Invalid decision."
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "outcome": {
                      "type": "string",
                      "enum": [
                        "not_found"
                      ]
                    }
                  },
                  "required": [
                    "outcome"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "outcome": {
                      "type": "string",
                      "enum": [
                        "already_decided"
                      ]
                    }
                  },
                  "required": [
                    "outcome"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}